Share some type define between pruntime and poller
